K12/requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.18
django-phonenumber-field==7.0.2
djangorestframework==3.12.4
drf-yasg==1.21.4
flake8==5.0.4
flake8-isort==5.0.0
gunicorn==20.0.4
idna==3.4
importlib-metadata==4.2.0
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mccabe==0.7.0
packaging==23.0
phonenumberslite==8.13.6
Pillow==9.4.0
psycopg2-binary==2.8.6
pycodestyle==2.9.1
pyflakes==2.5.0
python-dotenv==0.21.0
pytz==2022.6
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sorl-thumbnail==12.9.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
typing-extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.13
zipp==3.10.0
